chore: setup basic webserver
Set up a basic webserver with a route for GET /ping/ which returns 200
OK and an OK body in raw bytes.

The basic server has a couple flags:
- addr to define a running port
- debug to enable more verbose logging

The server uses structured logging through slog, with a JSON format.

TLS is enabled by default, this may be a bad choice for people who want
to run the app locally, but for now we'll use it. Instructions to
generate the key in the readme file.