feat: add ctrl-b reopen previous task
Once as task is closed, store the reference to the task in the browsers
session. Then when ctrl-b is invoked, reopen the task.

This unlocks a really helpful pattern - ability to jump between tasks.
If you need to reference a task while writing one, save the current
task, navigate to the task you're after (maybe an easymotion movement or
usisng the filter), read what you need, or update the task, then press
ctrl-b to jump back to the task you were at originally. Now you can
ctrl-b to jump back to the task for reference and again back to the
original.

If you end up closing a task, and didn't mean to, ctrl-b will reopen it.