chore: support new changelog syntax
Existing changelog syntax, which was little more than a rule, and a list
of timestamps was going to get harder to maintain over time, and didn't
have semantic meaning in the scheme of markdown.

This new approach, which uses definition list syntax allows for easier
regex matching, replacing, and for the changelog to exist anywhere in
the contents of the file, not just the last thing.